[FUSEQE-3296] Refactor package names
