+++ rest-tests/src/test/resources/features/irc.feature
@@ -16,12 +16,11 @@ Feature: Integration - IRC
     Given deploy ActiveMQ broker
       And create ActiveMQ connection
       And create ActiveMQ "subscribe" action step with destination type "queue" and destination name "irc-input"
-    # todo rephrase
-      And change datashape of previous step to "out" direction, "JSON_INSTANCE" type with specification '{"header":"text", "messageContent":"text"}'
+      And change "out" datashape of previous step to "JSON_INSTANCE" type with specification '{"header":"text", "messageContent":"text"}'
       And start mapper definition with name: "mapping 1"
       And MAP using Step 1 and field "/messageContent" to "/response/body"
       And create IRC "sendmsg" step with nickname "syndesis-publish" and channels "#spam1,#spam2"
-      And change datashape of previous step to "in" direction, "XML_INSTANCE" type with specification '<?xml version="1.0" encoding="UTF-8"?><response><body>message</body></response>'
+      And change "in" datashape of previous step to "XML_INSTANCE" type with specification '<?xml version="1.0" encoding="UTF-8"?><response><body>message</body></response>'
     When create integration with name: "AMQ-IRC"
     Then wait for integration with name: "AMQ-IRC" to become active
     When connect IRC listener to channels "#spam1,#spam2"
